requirements.txt
flask
flask-cors
python-dotenv
gunicorn==21.2.0

